use concerns to DRY DocumentsControllers
