+++ app/controllers/dashboard/base_controller.rb
@@ -46,18 +46,20 @@ class Dashboard::BaseController < ApplicationController
   end
 
   def protocol_authorizer_view
-    @authorization = ProtocolAuthorizer.new(@protocol, current_user)
+    @authorization      = ProtocolAuthorizer.new(@protocol, current_user)
+    @permission_to_view = @authorization.can_view?
 
     # Admins should be able to view too
-    unless @authorization.can_view? || @admin
+    unless @permission_to_view || @admin
       authorization_error('You are not allowed to access this protocol.')
     end
   end
 
   def protocol_authorizer_edit
-    @authorization = ProtocolAuthorizer.new(@protocol, current_user)
+    @authorization      = ProtocolAuthorizer.new(@protocol, current_user)
+    @permission_to_edit = @authorization.can_edit?
 
-    unless @authorization.can_edit? || @admin
+    unless @permission_to_edit || @admin
       authorization_error('You are not allowed to edit this protocol.')
     end
   end
