+++ app/controllers/documents_controller.rb
@@ -21,87 +21,6 @@
 class DocumentsController < ApplicationController
   before_action :initialize_service_request
   before_action :authorize_identity
-  before_action :find_document,             only: [:edit, :update, :destroy]
-  before_action :find_protocol,             only: [:index, :new, :create, :update]
 
-  def index
-    @documents = @protocol.documents.eager_load(:organizations)
-    
-    respond_to :json
-  end
-
-  def new
-    @document = @protocol.documents.new
-
-    respond_to :js
-  end
-
-  def create
-    @document = @protocol.documents.new(document_params)
-
-    if @document.save
-      assign_organization_access
-
-      flash.now[:success] = t(:documents)[:created]
-    else
-      @errors = @document.errors
-    end
-
-    respond_to :js
-  end
-
-  def edit
-    respond_to :js
-  end
-
-  def update
-    if @document.update_attributes(document_params)
-      assign_organization_access
-
-      flash.now[:success] = t(:documents)[:updated]
-    else
-      @errors = @document.errors
-    end
-
-    respond_to :js
-  end
-
-  def destroy
-    DocumentRemover.new(params[:id])
-
-    flash.now[:success] = t(:documents)[:destroyed]
-
-    respond_to :js
-  end
-
-  private
-
-  def document_params
-    params.require(:document).permit(
-      :document,
-      :doc_type,
-      :doc_type_other,
-      :sub_service_requests,
-      :protocol_id)
-  end
-
-  def find_document
-    @document = Document.find(params[:id])
-  end
-
-  def find_protocol
-    if @document
-      @protocol = @document.protocol
-    elsif @service_request
-      @protocol = @service_request.protocol
-    elsif params[:protocol_id]
-      @protocol = Protocol.find(params[:protocol_id])
-    else
-      @protocol = Protocol.find(document_params[:protocol_id])
-    end
-  end
-
-  def assign_organization_access
-    @document.sub_service_requests = @protocol.sub_service_requests.where(organization_id: params[:org_ids])
-  end
+  include DocumentsControllerShared
 end
